move interner types into `bevy_utils`
