+++ crates/bevy_ecs/Cargo.toml
@@ -26,8 +26,6 @@ fixedbitset = "0.4"
 fxhash = "0.2"
 downcast-rs = "1.2"
 serde = "1"
-parking_lot = "0.12"
-indexmap = "1.9.1"
 
 [dev-dependencies]
 rand = "0.8"
